Removed widget package.  More cleanup.  Purging.
