+++ app/src/main/java/tgd/mindless/drone/weatherwidgetnumberone/redux/Box.java
@@ -1,4 +1,4 @@
-package tgd.mindless.drone.weatherwidgetnumberone.redux.widget;
+package tgd.mindless.drone.weatherwidgetnumberone.redux;
 
 class Box {
     float left;
