+++ app/src/main/java/tgd/mindless/drone/weatherwidgetnumberone/redux/Ranges.java
@@ -1,0 +1,16 @@
+package tgd.mindless.drone.weatherwidgetnumberone.redux;
+
+class Ranges {
+
+    class Range {
+        float max;
+        float min;
+    }
+
+    Range temperature;
+    float temperatureMaxTime;
+
+    Ranges() {
+
+    }
+}
